moved some stuff out of DBConnection::deleteObj into DBAdaptro::deleteObj that could have resulted in errors in other (non-core) DBAdaptors that inherited from DBConnection

+++ modules/Bio/EnsEMBL/DBSQL/DBConnection.pm
@@ -521,23 +521,22 @@ sub get_db_adaptor {
 sub deleteObj {
   my $self = shift;
   
-#  print STDERR "DBConnection::deleteObj : Breaking circular references:\n";
+  #print STDERR "DBConnection::deleteObj : Breaking circular references:\n";
+  
+  if(exists($self->{'_adaptors'})) {
+    foreach my $adaptor_name (keys %{$self->{'_adaptors'}}) {
+      my $adaptor = $self->{'_adaptors'}->{$adaptor_name};
 
-    foreach my $key (qw{ current_objects _adaptors }) {
-      foreach my $adaptor_name (keys %{$self->{$key}}) {
-
-        my $adaptor = $self->{$key}->{$adaptor_name};
-
-        #call each of the adaptor deleteObj methods
-        if($adaptor && $adaptor->can('deleteObj')) {
-          #print STDERR "\t\tdeleting adaptor\n";
-          $adaptor->deleteObj();
-        }
-
-        #break dbadaptor -> object adaptor references
-        $self->{$key}->{$adaptor_name} = undef;
+      #call each of the adaptor deleteObj methods
+      if($adaptor && $adaptor->can('deleteObj')) {
+        #print STDERR "\t\tdeleting adaptor\n";
+        $adaptor->deleteObj();
       }
+
+      #break dbadaptor -> object adaptor references
+      delete $self->{'_adaptors'}->{$adaptor_name};
     }
+  }
 
   #print STDERR "Cleaning up attached databases\n";
